pyproject.toml
[project]
name = "raginit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb==0.5.0",
    "langchain==0.2.2",
    "langchain-community==0.2.3",
    "langchain-openai==0.1.8",
    "onnxruntime==1.17.1",
        "openai>=1.50.0",
    "python-dotenv==1.0.1",
    "tiktoken==0.7.0",
    "unstructured[md]==0.14.4",
]
